r/role_definition: refactoring to match the other state upgraders
